Se agrega unit test que testea getCRCPrices
